Remove tarball check when creating a new-release review
We are moving this script to be executed in post pipeline for
openstack/releases project and it may happen that the tarball
is not ready when running the job. In that case we prefer to
create the review and recheck if needed that getting an error
in the script and not creating the review.

Change-Id: Ic6be5fef273f6292bb0a4cb5b31e774d74409dd2