Make rdo_release_review stateless
In previous changes rdo_release_review has been modified
to support tag updates in rdoinfo as trigger to new releases
reviews. With this change, statefulness is not longer
required in rdo_release_review. This patch makes rdo_release_review
stateless and more suitable to be used in CI jobs in post
pipelines for openstack/releases and rdoinfo.

Change-Id: I103883d1e538baad2c2aa4631231be47b1990df0