Add update_uc command to rdoutils
As part of the rdoinfo reorganization, we need to update update-uc.py
[1] to use update_tag method in rdoutils.rdoinfo in releng to write new
tags in rdoinfo.

Additionally, this patch is moving update-uc.py from rdoinfo [1] to
releng so that we can make rdoinfo to contain only data itself and not
python code that should be moved to distroinfo or CI tooling repos, not
rdoinfo itself.

[1] https://github.com/redhat-openstack/rdoinfo/blob/master/update-uc.py

Change-Id: I21562817f03a9f62dfffda7106d9e954977bf2bc