Add parameter in rdo_release_review to override the rdoinfo tag
Currently, rdo_release_review always asumes that the tag name in rdoinfo
for a release is the release name. However, during the release
preparation we use a different one <release>-uc.

Allowing to specify the tag in the rdo_release_review allows us to
enable this automation before creating the new release train in rdoinfo.

Change-Id: I2e7face34054609c10b96c06751c750bf9c358cd