Reorganize args for ease of importing
Initially, I chose to use the args of parser.parse_args()
directly in functions and not passed as function parameters
to keep code readability. But this makes 'import' impossible.
As now we need to use this script as a module to be able to
pusblish a report in dashboards.r.o, I passed those args
to function.

Plus, we add the to_UpperConstraint::to_list() method.

Change-Id: Id4e49047cd5bb8db760b5c4f66a3ae93bc2c9f24