Remove only the CONSTRAINTS_FILE
There are case CONSTRAINTS file is in the same line
as 'deps ='. With the current sed command, we remove
the whole line, including the 'deps =' which makes fail
tox command.
This patch adds another sed command to handle this case.

Also, we edit the regexp to include 'UPPER_' and 'TOX_'.

Change-Id: I8dfcd9d44a398bb5eb548ba392f6a487cffb5732